requirements.txt
PyPDF2==3.0.1
langchain>=0.0.200
langchain-openai>=0.0.4
tiktoken>=0.3.3
faiss-gpu==1.7.4
openai>=0.27.8
tenacity>=8.2.2
psycopg2>=2.9.7
SQLAlchemy>=2.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0